+++ src/lib/hooks/useSiteSearch.tsx
@@ -1,14 +1,17 @@
 import {useEffect, useState} from "react";
 import axios from "axios";
 
-export const useSiteSearch = (query, limit = 9999) => {
+export const useSiteSearch = (query, limit = 9999, callback = null) => {
   const [results, setResults] = useState([]);
 
   useEffect(() => {
     const fetchResults = async () => {
       const url = `/api/site-search?q=${query}`
       await axios.get(url)
-        .then(result => setResults(result.data))
+        .then(result => {
+          setResults(result.data);
+          callback?.(result.data)
+        })
         .catch(e => console.error(e.message))
     }
     if (query) fetchResults();
